==> 01_read_data.ipynb <==
## Branje podatkov

import pandas as pd
import os.path

Preberemo podatke z intenzitetani izražanja genov, eksperimentalno pridobljenih z DNA mirkomrežami:

df = pd.read_csv(os.path.join('data','data.txt'), comment="#", sep="\t")

Kaj podatki vsebujejo?

df.columns

Stolpca `'1 Avg (log2)'` in `'2 Avg (log2)'` vsebujeta povprečne intenzitete treh kontrol in treh vzorcev z utišanjem proteina ESRP2. Zadnjih 6 stolpcev predstavlja izmerjene intenzitete pri vseh 6 vzorchih (3 x kontrola, 3 x knockdown). Eksperiment je opisan v [članku](https://www.mdpi.com/1422-0067/22/14/7477).

Zanimata nas zgolj stolpca s povprečji skupaj z imeni genov, ki jih uporablja tudi GEM:

df1 = df[['Gene Symbol', '1 Avg (log2)']]
df2 = df[['Gene Symbol', '2 Avg (log2)']] 

Podatke si shranimo v ločeni datoteki:

df1.columns=['gene', 'value']
df2.columns=['gene', 'value']

df1.to_csv(os.path.join('data','data_control.txt'), index=False)
df2.to_csv(os.path.join('data','data_kd.txt'), index=False)
