==> 03_flux_sampling.ipynb <==
import pandas as pd
import cobra

from cobra.io import load_model

import numpy as np
import pickle

import os.path

## Osnovne nastavitve

require_biomass = True
folder = os.path.join('models','biomass') if require_biomass else os.path.join('models','no_biomass')
folder_samples = os.path.join('samples','biomass') if require_biomass else os.path.join('samples','biomass')

## Preberi modele

model_control = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_control.xml'))
model_kd = cobra.io.read_sbml_model(os.path.join(f'{folder}','model_kd.xml'))

Pridobi reakcije v posameznem modelu in jih zapiši kot množica (`set`).

reacts_control = {r.id for r in model_control.reactions}
reacts_kd = {r.id for r in model_kd.reactions}

## Osnovne analize

### Jaccardov indeks

Jaccardov indeks/koeficient je statistična mera podobnosti dveh populacij/vzorcev.

Izračuna se kot razmerje med številom enakih reakcij v obeh vzorcih (presek množic) in številom vseh različnih reakcij v vzorcih (unija množic):
$$J=\frac{|S_1 \cap S_2|}{|S_1 \cup S_2|}$$

Ker je moč preseka v najboljšem primeru enaka moči unije, se indeks nahaja na intervalu $[0,1]$. Višji Jaccardov indeks pomeni večjo podobnost med opazovanima vzorcema.

J = len(reacts_control & reacts_kd)/len(reacts_control | reacts_kd)
J

## Summary

model_control.summary()

model_kd.summary()

## Vzorčenje metabolnih pretokov

Uporabili bomo funkcijo sample paketa COBRApy ([povezava](https://cobrapy.readthedocs.io/en/latest/sampling.html)).

from cobra.sampling import sample

Podamo model, število vzorcev in metodo vzorčenja. 

**POZOR: Časovno potratno! Podatke lahko prebereš iz datoteke (spodaj).**

s_control = sample(model_control, 1000, method="achr") 

**POZOR: Časovno potratno! Podatke lahko prebereš iz datoteke (spodaj).**

s_kd = sample(model_kd, 1000, method="achr")

s_control.to_csv(os.path.join(f'{folder_samples}','samples_control.csv'), index=False)

s_kd.to_csv(os.path.join(f'{folder_samples}','samples_kd.csv')', index=False)

**Branje iz datotek (če ste časovne potratne korake izpustili)**

s_control = pd.read_csv(os.path.join(f'{folder_samples}','samples_control.csv'))
s_kd = pd.read_csv(os.path.join(f'{folder_samples}','samples_kd.csv'))

s_control

s_kd
